Organise firebase classes
