+++ app/src/main/kotlin/com/jpaya/englishisfun/data/firebase/di/FirebaseModule.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.jpaya.englishisfun.firestore
+package com.jpaya.englishisfun.data.firebase.di
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.ktx.firestore
