+++ app/src/main/kotlin/com/jpaya/englishisfun/idioms/data/network/NetworkDataSource.kt
@@ -16,7 +16,7 @@
 
 package com.jpaya.englishisfun.idioms.data.network
 
-import com.jpaya.englishisfun.firestore.FireStoreClient
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.jpaya.englishisfun.idioms.mapper.toDomain
 import com.jpaya.englishisfun.idioms.domain.Idiom
 import javax.inject.Inject
