+++ app/src/main/kotlin/com/jpaya/englishisfun/phrasals/data/network/NetworkDataSource.kt
@@ -16,7 +16,7 @@
 
 package com.jpaya.englishisfun.phrasals.data.network
 
-import com.jpaya.englishisfun.firestore.FireStoreClient
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.jpaya.englishisfun.phrasals.domain.Phrasal
 import com.jpaya.englishisfun.phrasals.mapper.toDomain
 import javax.inject.Inject
