+++ app/src/main/kotlin/com/jpaya/englishisfun/suggestions/data/network/NetworkDataSource.kt
@@ -16,7 +16,7 @@
 
 package com.jpaya.englishisfun.suggestions.data.network
 
-import com.jpaya.englishisfun.firestore.FireStoreClient
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.jpaya.englishisfun.suggestions.domain.Suggestions
 import com.jpaya.englishisfun.suggestions.mapper.toNetwork
 import javax.inject.Inject
