+++ app/src/test/kotlin/com/jpaya/englishisfun/data/firebase/FireStoreClientTest.kt
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com.jpaya.englishisfun.firestore
+package com.jpaya.englishisfun.data.firebase
 
 import com.google.android.gms.tasks.Task
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.FirebaseFirestore
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
