+++ app/src/test/kotlin/com/jpaya/englishisfun/data/firebase/di/FirebaseModuleTest.kt
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.jpaya.englishisfun.firestore
+package com.jpaya.englishisfun.data.firebase.di
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
+import com.jpaya.englishisfun.data.firebase.di.FirebaseModule
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
