+++ app/src/test/kotlin/com/jpaya/englishisfun/irregulars/data/network/NetworkDataSourceTest.kt
@@ -16,7 +16,7 @@
 
 package com.jpaya.englishisfun.irregulars.data.network
 
-import com.jpaya.englishisfun.firestore.FireStoreClient
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.jpaya.englishisfun.irregulars.data.network.model.IrregularNetworkItem
 import com.jpaya.englishisfun.irregulars.data.network.model.IrregularsResponse
 import com.jpaya.englishisfun.irregulars.domain.Irregular
