+++ app/src/test/kotlin/com/jpaya/englishisfun/statives/data/network/NetworkDataSourceTest.kt
@@ -16,7 +16,7 @@
 
 package com.jpaya.englishisfun.statives.data.network
 
-import com.jpaya.englishisfun.firestore.FireStoreClient
+import com.jpaya.englishisfun.data.firebase.FireStoreClient
 import com.jpaya.englishisfun.statives.data.network.model.StativeNetworkItem
 import com.jpaya.englishisfun.statives.data.network.model.StativesResponse
 import com.jpaya.englishisfun.statives.domain.Stative
